refactor(ui): separate Electrical metering panel, restore release cumuls panel (#376)
Per review: the Consommation energie panel is byte-identical to the
release again (no live tiles inside, original render condition). The
live P/U/I/PF values move to a dedicated Electrical metering panel
rendered below it, visible only when at least one of the four measures
is bound.
